% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{encrypt_password_v10}
\alias{encrypt_password_v10}
\title{Instagram Password Encryption}
\usage{
encrypt_password_v10(key_id, public_key, password)
}
\arguments{
\item{key_id}{The key ID for the encryption}

\item{public_key}{The public key for the encryption}

\item{password}{The password to encrypt}
}
\value{
The encrypted password
}
\description{
Encrypts an Instagram password using the rinstagram ML microservice
(\code{POST /encrypt}), which implements Instagram's v10 AES-GCM + NaCl
SealedBox encryption protocol.
}
