% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_csrf_token}
\alias{generate_csrf_token}
\title{CSRF token generation function}
\usage{
generate_csrf_token()
}
\value{
A random CSRF token
}
\description{
A function to generate a random CSRF token for Instagram requests
}
\examples{
my_csrf_token <- generate_csrf_token()
}
