% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_random_user_agent}
\alias{get_random_user_agent}
\title{User-Agent retrieval function}
\usage{
get_random_user_agent()
}
\value{
A random User-Agent string
}
\description{
A function to retrieve a random User-Agent string
}
\examples{
my_random_user_agent <- get_random_user_agent()
}
