% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{handle_login}
\alias{handle_login}
\title{Instagram Login Handler}
\usage{
handle_login(username, password)
}
\arguments{
\item{username}{The username to login with}

\item{password}{The password to login with}
}
\value{
A list containing the response data, the InstagramSession object, and the cookies
}
\description{
A function to handle a login request to Instagram
}
