% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_profile.R
\name{handle_web_profile_request}
\alias{handle_web_profile_request}
\title{Instagram Web Profile Request Handler}
\usage{
handle_web_profile_request(username)
}
\arguments{
\item{username}{The username of the Instagram profile to fetch}
}
\value{
An InstagramProfileWebProfile object
}
\description{
A function to handle a web profile request for a given Instagram profile
}
\examples{
osamason_user_object <- handle_web_profile_request("osamason")
}
