% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_profile.R
\name{handle_web_profile_requests}
\alias{handle_web_profile_requests}
\title{Instagram Web Profile Request Handler}
\usage{
handle_web_profile_requests(usernames)
}
\arguments{
\item{usernames}{A list of usernames to fetch}
}
\value{
A list of InstagramProfileWebProfile objects
}
\description{
A function to handle multiple web profile requests for a list of Instagram profiles
}
\examples{
user_objects <- handle_web_profile_requests(c("osamason", "instagram"))
}
