% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{InstagramProfile-class}
\alias{InstagramProfile-class}
\title{InstagramProfile Data Class}
\description{
A class to store Instagram profile data
}
\section{Fields}{

\describe{
\item{\code{username}}{The username of the Instagram profile}

\item{\code{followerCount}}{The number of followers the Instagram profile has}

\item{\code{followingCount}}{The number of accounts the Instagram profile is following}

\item{\code{posts_count}}{The number of posts the Instagram profile has}

\item{\code{posts_likes}}{A optional list of the number of likes each post has}

\item{\code{posts_comments}}{A optional list of the number of comments each post has}

\item{\code{posts_dates}}{A optional list of the dates each post was posted}
}}

