% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{InstagramProfileWebProfile-class}
\alias{InstagramProfileWebProfile-class}
\title{InstagramProfileWebProfile Data Class}
\description{
A class to store raw Instagram profile data.
}
\section{Fields}{

\describe{
\item{\code{username}}{The username of the Instagram profile.}

\item{\code{followerCount}}{The number of followers the Instagram profile has.}

\item{\code{followingCount}}{The number of accounts the Instagram profile is following.}

\item{\code{posts}}{The number of posts the Instagram profile has.}
}}

\examples{
new_profile <- new("InstagramProfileWebProfile", username = "osamason", followerCount = 1000, followingCount = 100, posts = 50)
}
