% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{class}
\name{InstagramSession-class}
\alias{InstagramSession-class}
\title{InstagramSession Data Class}
\description{
A class to store Instagram session data
}
\section{Fields}{

\describe{
\item{\code{csrf}}{The CSRF token for the session}

\item{\code{user_agent}}{The User-Agent string for the session}

\item{\code{session_id}}{The session ID for the session (optional)}

\item{\code{user_id}}{The user ID for the session (optional)}
}}

