% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login_request}
\alias{login_request}
\title{Instagram Login Request}
\usage{
login_request(instagram_session, username, password, key_id, public_key)
}
\arguments{
\item{instagram_session}{An InstagramSession object containing the user agent and CSRF token}

\item{username}{The username to login with}

\item{password}{The password to login with}

\item{key_id}{The key ID for the encryption}

\item{public_key}{The public key for the encryption}
}
\value{
A list containing the response data, the InstagramSession object, and the cookies
}
\description{
A function to make a login request to Instagram
}
