% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_instagram_profile}
\alias{save_instagram_profile}
\title{Save Instagram profile data to CSV}
\usage{
save_instagram_profile(instagram_profile, filePath = "data/profiles.csv")
}
\arguments{
\item{instagram_profile}{An InstagramProfile object}

\item{filePath}{The file path to save the data to}
}
\description{
A function to save Instagram profile data to a CSV file
}
