% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_profiles_to_csv}
\alias{save_profiles_to_csv}
\title{CSV saving function}
\usage{
save_profiles_to_csv(profiles, filepath = "instagram_profiles.csv")
}
\arguments{
\item{profiles}{A list of InstagramProfileRaw objects}

\item{filepath}{The file path to save the CSV file to}
}
\description{
A function to save Instagram profile data to a CSV file
}
\examples{
save_profiles_to_csv(profiles, "/path/to/data.csv")
save_profiles_to_csv(profiles)
}
