% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveInstagramProfile}
\alias{saveInstagramProfile}
\title{Save Instagram profile data to CSV}
\usage{
saveInstagramProfile(instagramProfile, filePath = "data/profiles.csv")
}
\arguments{
\item{instagramProfile}{An InstagramProfile object}

\item{filePath}{The file path to save the data to}
}
\description{
A function to save Instagram profile data to a CSV file
}
