% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{shared_data_request}
\alias{shared_data_request}
\title{Instagram Shared Data Request}
\usage{
shared_data_request(instagram_session)
}
\arguments{
\item{instagram_session}{An InstagramSession object containing the user agent and CSRF token}
}
\value{
A list containing the shared data and the InstagramSession object
}
\description{
A function to make a request to the Instagram shared data API to get the encryption keys
}
