% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_profile.R
\name{web_profile_request}
\alias{web_profile_request}
\title{Instagram Web Profile Request}
\usage{
web_profile_request(csrf, user_agent, username)
}
\arguments{
\item{csrf}{The CSRF token to use in the request}

\item{user_agent}{The User-Agent string to use in the request}

\item{username}{The username of the Instagram profile to fetch}
}
\value{
The raw JSON response from the Instagram web profile API
}
\description{
A function to make a request to the Instagram web profile API
}
\examples{
data <- web_profile_request("AoH9uiQnUrg0tfceSN9dYbphGEQN6N59", "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:109.0) Gecko/20100101 Firefox/115.0", "osamason")
}
