% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare}
\alias{compare}
\title{Compare Instagram profile data}
\usage{
compare(file_path, path_to_save = NA)
}
\arguments{
\item{file_path}{The file path to the CSV file}

\item{path_to_save}{The optional file path to save the summary data to}
}
\value{
A data frame containing the summary of the data
}
\description{
A function to compare Instagram profile data from a CSV file
}
\examples{
instagram_data_comparison <- compare("path_to_your_instagram_data.csv")
instagram_data_comparison <- compare("path_to_your_instagram_data.csv", "path_to_save_summary.csv")
}
