% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lscrape}
\alias{lscrape}
\title{Scrape Instagram profiles (with session) [lscrape -> logged in scraping]}
\usage{
lscrape(usernames, profile_username, profile_password, file_path)
}
\arguments{
\item{usernames}{A list of usernames to scrape}

\item{profile_username}{The username of the Instagram account to use for scraping}

\item{profile_password}{The password of the Instagram account to use for scraping}

\item{file_path}{The optional file path to save the scraped data to (default: "data/profiles.csv")}
}
\value{
A data frame containing the scraped data (username, followers, following, post count) or NULL if an error occurred
}
\description{
A function to scrape Instagram profiles with a session
You need an account to use this function
Do NOT use your own account for scraping as it might get rate limited
Use a dummy account or a throwaway account
}
\examples{
scraped_data <- lscrape(c("osamason", "praiseche"), "mytestaccount", "mypassword", "path/to/profiles.csv")
}
