% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{web_profile_request}
\alias{web_profile_request}
\title{Instagram Web Profile Request}
\usage{
web_profile_request(instagram_session, username)
}
\arguments{
\item{instagram_session}{An InstagramSession object containing the user agent and CSRF token}

\item{username}{The username of the Instagram profile to fetch}
}
\value{
The raw JSON response from the Instagram web profile API
}
\description{
A function to make a request to the Instagram web profile API
}
